+++ Database/constants.lua
@@ -262,6 +262,16 @@ Constants.SpecIDs = {
 	73,
 }
 
+Constants.SatedDebuffs = {
+	[57723] = true,
+	[57724] = true,
+	[80354] = true,
+	[95809] = true,
+	[160455] = true,
+	[264689] = true,
+	[390435] = true,
+}
+
 Constants.ClassSecondaryPower = {
 	["DEATHKNIGHT"] = {
 		resourceKind = "runes",
@@ -557,6 +567,8 @@ Constants.FakeAuras = {
 	[451234] = true, -- Voidwrath 6
 	[34433] = true, -- Shadowfiend 6
 	[1280137] = true, -- Mindbender 12
+	[450193] = true, -- Entropic Rift
+	[449880] = true, -- Void Heart
 
 	-- SHAMAN
 	[5394] = true, -- Healing Stream Totem 15