+++ Modules/glow.lua
@@ -46,6 +46,30 @@ function SCM:StartCustomGlow(child)
 			"SCM",
 			1
 		)
+
+		-- Why do I have to do this?
+		local glowFrame = child["_PixelGlowSCM"]
+		if glowFrame then
+			glowFrame:ClearAllPoints()
+			glowFrame:SetPoint("TOPLEFT", child, "TOPLEFT", -glowTypeOptions.xOffset, glowTypeOptions.yOffset)
+			glowFrame:SetPoint("BOTTOMRIGHT", child, "BOTTOMRIGHT", glowTypeOptions.xOffset, -glowTypeOptions.yOffset)
+
+			for _, texture in pairs(glowFrame.textures) do
+				texture:SetTexelSnappingBias(0)
+				texture:SetSnapToPixelGrid(false)
+
+				if glowTypeOptions.border then
+					texture:SetBlendMode("ADD")
+				end
+			end
+
+			for _, mask in pairs(glowFrame.masks) do
+				mask:SetTexelSnappingBias(0)
+				mask:SetSnapToPixelGrid(false)
+			end
+		end
+	elseif options.glowType == "Button" then
+		LibCustomGlow.ButtonGlow_Start(child, color, glowTypeOptions.frequency)
 	end
 
 	activeGlows[child] = true
@@ -58,6 +82,8 @@ function SCM:StopCustomGlow(child)
 		LibCustomGlow.AutoCastGlow_Stop(child, "SCM")
 	elseif child.SCMGlow == "Pixel" then
 		LibCustomGlow.PixelGlow_Stop(child, "SCM")
+	elseif child.SCMGlow == "Button" then
+		LibCustomGlow.ButtonGlow_Stop(child)
 	end
 
 	child.SCMGlow = nil
@@ -92,7 +118,7 @@ end
 
 function SCM:RestoreBlizzardGlows()
 	local options = self.db.profile.options
-	for _, viewerName in ipairs({"EssentialCooldownViewer", "UtilityCooldownViewer", "BuffIconCooldownViewer"}) do
+	for _, viewerName in ipairs({ "EssentialCooldownViewer", "UtilityCooldownViewer", "BuffIconCooldownViewer" }) do
 		local viewer = _G[viewerName]
 		if viewer then
 			for _, child in ipairs({ viewer:GetChildren() }) do