+++ Modules/profiles.lua
@@ -11,6 +11,7 @@ local GLOBAL_CUSTOM_CONFIG_KEYS = {
 	"itemConfig",
 	"slotConfig",
 	"timerConfig",
+	"bloodlustConfig",
 }
 local PROFILE_OPTION_SECTION_KEYS = {
 	resourceBar = true,
@@ -299,6 +300,7 @@ function SCM:ImportProfile(profileName, importString)
 		profileName = SCM.db:GetCurrentProfile()
 	end
 
+	SCM.importingProfile = true
 	SCM.db:SetProfile(profileName)
 
 	local db = self.db.profile
@@ -353,12 +355,13 @@ function SCM:ImportProfile(profileName, importString)
 	end
 
 	self.db.profile.options = options
+	SCM.db:RegisterDefaults(SCM.DefaultDB)
 
 	SCM.appliedOptions = nil
 	SCM:ApplyOptions()
-	SCM.db:RegisterDefaults(SCM.DefaultDB)
 
 	SCM.RefreshCooldownViewerData(true)
+	SCM.importingProfile = nil
 end
 
 function SCM:ImportGlobalSettings(importString)