+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-### Version 1.5.30 [Alpha]
+### Version 1.5.34 [Alpha]
 
 Additions:
+- Add more presets (potions, racials, etc)
+- Add option for a "Smooth" resource bar
 
 Fixes:
-- Fix the background anchoring for "Bars Only" buff bars
+- Fix 12.1 lua errors
+- Hopefully fix stuck cooldowns
+- Custom icon cooldown and charge display improvements and fixes
+- Fix the edge and swipe for abilities with charges and an active buff
+- Go back to the old cooldown frame anchoring (hopfully nothing breaks)
+- Fix lua errors when importing profiles
+- Make sure "always show" only works when "Disable "Hide While Inactive"" is enabled
+- Clean up resource bar code
+- Make sure the border is disabled when it should be